+++ src/changelog.ts
@@ -1,15 +1,15 @@
 import dotenv from 'dotenv';
+import { TaskTree } from 'tasktree-cli';
 import Reader from './io/reader';
 import Writer from './io/writer';
 import Provider, { ProviderName } from './providers/provider';
 import GitHubProvider from './providers/github-provider';
 import Config, { ConfigOptions } from './entities/config';
 import Package from './entities/package';
-import Process from './utils/process';
 
 dotenv.config();
 
-const $process = Process.getInstance();
+const $tasks = TaskTree.tree();
 
 export default class Changelog {
     private config: Config;
@@ -17,11 +17,10 @@ export default class Changelog {
     private reader: Reader | undefined;
 
     public constructor(options?: ConfigOptions) {
-        const task = $process.task('Reading configuration files');
+        const task = $tasks.add('Reading configuration files');
 
         this.config = new Config(options);
         this.pkg = new Package();
-        // this.writer = new Writer();
         this.reader = this.getReader();
 
         if (!this.reader) task.fail(`Provider or Reader is not specified (${this.config.provider})`);
@@ -30,13 +29,14 @@ export default class Changelog {
     }
 
     public async generate(): Promise<void> {
-        const { config, pkg, reader } = this;
+        const { reader } = this;
 
         if (reader) {
             const state = await reader.read();
+            const writer = new Writer(this.pkg);
 
-            await state.modify(config);
-            await Writer.write(state, pkg);
+            await state.modify(this.config);
+            await writer.write(state);
         }
     }