+++ src/utils/types.ts
@@ -1,15 +1,11 @@
 export type OptionValue = string | string[] | boolean | undefined;
 
-export interface ReadonlyArray<T> extends Array<T> {
-    readonly [n: number]: T;
-}
-
 export interface Constructable<T, C> {
-    new(context: C): T;
+    new (context: C): T;
 }
 
 export interface Importable<T, C> {
-    default: Constructable<T, C>
+    default: Constructable<T, C>;
 }
 
 export interface Option {